yawkonaa

+++ src/main/java/com/roi/employeee/model/Employee.java
@@ -10,7 +10,7 @@ public class Employee {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private String email;
@@ -18,18 +18,18 @@ public class Employee {
     public Employee() {}
 
 
-    public Employee(Integer id, String name, String email) {
+    public Employee(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
